refactor(agent): extract startup preflight module (#226)
