+++ crates/pi-coding-agent/src/startup_preflight.rs
@@ -1,0 +1,48 @@
+use super::*;
+
+pub(crate) fn execute_startup_preflight(cli: &Cli) -> Result<bool> {
+    if cli.session_validate {
+        validate_session_file(cli)?;
+        return Ok(true);
+    }
+
+    if cli.channel_store_inspect.is_some() || cli.channel_store_repair.is_some() {
+        execute_channel_store_admin_command(cli)?;
+        return Ok(true);
+    }
+
+    if cli.event_webhook_ingest_file.is_some() {
+        validate_event_webhook_ingest_cli(cli)?;
+        let payload_file = cli
+            .event_webhook_ingest_file
+            .clone()
+            .ok_or_else(|| anyhow!("--event-webhook-ingest-file is required"))?;
+        let channel_ref = cli
+            .event_webhook_channel
+            .clone()
+            .ok_or_else(|| anyhow!("--event-webhook-channel is required"))?;
+        let event_webhook_secret = resolve_secret_from_cli_or_store_id(
+            cli,
+            cli.event_webhook_secret.as_deref(),
+            cli.event_webhook_secret_id.as_deref(),
+            "--event-webhook-secret-id",
+        )?;
+        ingest_webhook_immediate_event(&EventWebhookIngestConfig {
+            events_dir: cli.events_dir.clone(),
+            state_path: cli.events_state_path.clone(),
+            channel_ref,
+            payload_file,
+            prompt_prefix: cli.event_webhook_prompt_prefix.clone(),
+            debounce_key: cli.event_webhook_debounce_key.clone(),
+            debounce_window_seconds: cli.event_webhook_debounce_window_seconds,
+            signature: cli.event_webhook_signature.clone(),
+            timestamp: cli.event_webhook_timestamp.clone(),
+            secret: event_webhook_secret,
+            signature_algorithm: cli.event_webhook_signature_algorithm.map(Into::into),
+            signature_max_skew_seconds: cli.event_webhook_signature_max_skew_seconds,
+        })?;
+        return Ok(true);
+    }
+
+    Ok(false)
+}
